requirements.txt
django==3.2.6
djangorestframework==3.13.1
python-decouple==3.5
psycopg2==2.9.3